+++ src/main/java/com/plms/springframework/bean/factory/config/ConfigurableBeanFactory.java
@@ -14,4 +14,9 @@ public interface ConfigurableBeanFactory extends HierarchicalBeanFactory, Single
      * @param beanPostProcessor
      */
     void addBeanPostProcessor(BeanPostProcessor beanPostProcessor);
+
+    /**
+     * 销毁单例Bean
+     */
+    void destroySingletons();
 }