+++ src/main/java/com/plms/springframework/bean/factory/support/DefaultSingletonBeanRegistry.java
@@ -1,9 +1,12 @@
 package com.plms.springframework.bean.factory.support;
 
+import com.plms.springframework.bean.BeansException;
+import com.plms.springframework.bean.factory.DisposableBean;
 import com.plms.springframework.bean.factory.config.SingletonBeanRegistry;
 
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Set;
 
 /**
  * @Author bigboss
@@ -11,7 +14,9 @@ import java.util.Map;
  */
 public class DefaultSingletonBeanRegistry implements SingletonBeanRegistry {
 
-    private Map<String, Object> singletonObjects = new HashMap<>();
+    private final Map<String, Object> singletonObjects = new HashMap<>();
+
+    private final Map<String, DisposableBean> disposableBeans = new HashMap<>();
 
     @Override
     public Object getSingleton(String beanName) {
@@ -22,5 +27,19 @@ public class DefaultSingletonBeanRegistry implements SingletonBeanRegistry {
         singletonObjects.put(beanName, bean);
     }
 
+    public void registerDisposableBean(String beanName, DisposableBean bean) {
+        disposableBeans.put(beanName, bean);
+    }
 
+    public void destroySingletons() {
+        Set<String> beanNames = disposableBeans.keySet();
+        for (String beanName : beanNames) {
+            DisposableBean disposableBean = disposableBeans.remove(beanName);
+            try {
+                disposableBean.destroy();
+            } catch (Exception e) {
+                throw new BeansException("Destroy method on bean with name '" + beanName + "' throw an exception", e);
+            }
+        }
+    }
 }