+++ src/main/java/com/plms/springframework/bean/factory/support/DisposableBeanAdapter.java
@@ -1,0 +1,42 @@
+package com.plms.springframework.bean.factory.support;
+
+import cn.hutool.core.util.StrUtil;
+import com.plms.springframework.bean.BeansException;
+import com.plms.springframework.bean.factory.DisposableBean;
+import com.plms.springframework.bean.factory.config.BeanDefinition;
+
+import java.lang.reflect.InvocationTargetException;
+import java.lang.reflect.Method;
+
+/**
+ * @Author bigboss
+ * @Date 2021/11/7 15:53
+ */
+public class DisposableBeanAdapter implements DisposableBean {
+
+    private final Object bean;
+
+    private final String beanName;
+
+    private String destroyMethodName;
+
+    public DisposableBeanAdapter(Object bean, String beanName, BeanDefinition beanDefinition) {
+        this.bean = bean;
+        this.beanName = beanName;
+        this.destroyMethodName = beanDefinition.getDestroyMethodName();
+    }
+
+    @Override
+    public void destroy() throws Exception {
+        if (bean instanceof DisposableBean) {
+            ((DisposableBean) bean).destroy();
+        }
+        if (StrUtil.isNotEmpty(destroyMethodName) && !(bean instanceof DisposableBean && "destroy".equals(this.destroyMethodName))) {
+            Method method = bean.getClass().getMethod(destroyMethodName);
+            if (null == method) {
+                throw new BeansException("Couldn't find a destroy method named '" + destroyMethodName + "' on bean with name '" + beanName + "'");
+            }
+            method.invoke(bean);
+        }
+    }
+}