+++ src/test/java/com/plms/springframework/bean/Person.java
@@ -1,5 +1,7 @@
 package com.plms.springframework.bean;
 
+import com.plms.springframework.bean.factory.DisposableBean;
+import com.plms.springframework.bean.factory.InitializingBean;
 import lombok.Data;
 
 /**
@@ -8,11 +10,62 @@ import lombok.Data;
  */
 
 @Data
-public class Person {
+public class Person implements DisposableBean, InitializingBean {
 
     private String name;
 
     private int age;
 
     private Car car;
+
+    public void customInitMethod() {
+        System.out.println("I was born in the method named customInitMethod");
+    }
+
+    public void customDestroyMethod() {
+        System.out.println("I died in the method named customDestroyMethod");
+    }
+
+    @Override
+    public void afterPropertiesSet() throws BeansException {
+        System.out.println("I was born in the method named afterPropertiesSet");
+    }
+
+    @Override
+    public void destroy() throws Exception {
+        System.out.println("I died in the method named destroy");
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public int getAge() {
+        return age;
+    }
+
+    public void setAge(int age) {
+        this.age = age;
+    }
+
+    public Car getCar() {
+        return car;
+    }
+
+    public void setCar(Car car) {
+        this.car = car;
+    }
+
+    @Override
+    public String toString() {
+        return "Person{" +
+                "name='" + name + '\'' +
+                ", age=" + age +
+                ", car=" + car +
+                '}';
+    }
 }